feat(test-support): Allow reusing snapbox assertions
